update alumni's prior positions and sort by alphabetical order
