+++ src/utils/data/alumni.tsx
@@ -1,0 +1,340 @@
+import { InternInformation } from '../../types/types';
+
+export const INTERN_ALUMNI: Array<InternInformation> = [
+  {
+    Intern_ID: 2,
+    Name: 'Aditi Dass',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'State Street',
+  },
+  {
+    Intern_ID: 54,
+    Name: 'Ammar Ahmad',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 8,
+    Name: 'Anna Goncharova',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Twitter',
+  },
+  {
+    Intern_ID: 23,
+    Name: 'Anya Gilad',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Athenahealth',
+  },
+  {
+    Intern_ID: 40,
+    Name: 'Ashni Shah',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 3,
+    Name: 'Ben Getchell',
+    Prior_Position: 'Software Engineer Intern, Software Engineer',
+    Post_Internship: 'Software & Application Innovation Lab',
+  },
+  {
+    Intern_ID: 1,
+    Name: 'Chang Gao',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'University of Virginia > Microsoft',
+  },
+  {
+    Intern_ID: 18,
+    Name: 'Chen Garber',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Facebook',
+  },
+  {
+    Intern_ID: 51,
+    Name: 'Christopher Cho',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 41,
+    Name: 'Christopher Trinh',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 21,
+    Name: 'Daniel Valentine',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 35,
+    Name: 'David Shen',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 5,
+    Name: 'David Wang',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Amazon',
+  },
+  {
+    Intern_ID: 38,
+    Name: 'Efstathios Karatsiolis',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 4,
+    Name: 'Ekaterina Propokeva',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Amazon Web Services',
+  },
+  {
+    Intern_ID: 22,
+    Name: 'Esen Harris',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Accenture',
+  },
+  {
+    Intern_ID: 39,
+    Name: 'FanonX Rogers',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 31,
+    Name: 'Francis Zamora',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Twilio',
+  },
+  {
+    Intern_ID: 45,
+    Name: 'Gagandeep Kang',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Amazon',
+  },
+  {
+    Intern_ID: 29,
+    Name: 'Glenn Liem',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 59,
+    Name: 'Greg Frasco',
+    Prior_Position: 'Software Engineer',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 11,
+    Name: 'Ibrahim Shaikh',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'ITG',
+  },
+  {
+    Intern_ID: 56,
+    Name: 'Ira Globus-Harris',
+    Prior_Position: 'Software Engineer',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 6,
+    Name: 'Jacqueline You',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Boston University School of Medicine',
+  },
+  {
+    Intern_ID: 10,
+    Name: 'Johnson Lam',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Soroco',
+  },
+  {
+    Intern_ID: 16,
+    Name: 'Justin Chen',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Systems & Technology Research',
+  },
+  {
+    Intern_ID: 17,
+    Name: 'Kinan Dak Albab',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Current BU Ph.D. Student',
+  },
+  {
+    Intern_ID: 12,
+    Name: 'Linlan Chen',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Facebook',
+  },
+  {
+    Intern_ID: 55,
+    Name: 'Lucy Qin',
+    Prior_Position: 'Software Engineer',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 43,
+    Name: 'Marcella Hastings',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 36,
+    Name: 'Megan Fantes',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'iRobot',
+  },
+  {
+    Intern_ID: 52,
+    Name: 'Nicholas Goutermout',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 53,
+    Name: 'Nicholas Hall',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 32,
+    Name: 'Nicolas Alhaddad',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 28,
+    Name: 'Peter Flockhart',
+    Prior_Position: 'Software Engineer Intern, Software Engineer',
+    Post_Internship: 'Software & Application Innovation Lab',
+  },
+  {
+    Intern_ID: 19,
+    Name: 'Qian Mei',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Growald Family Fund',
+  },
+  {
+    Intern_ID: 30,
+    Name: 'Rachel Manzelli',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 49,
+    Name: 'Raj Vipani',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Galatea Associates',
+  },
+  {
+    Intern_ID: 7,
+    Name: 'Raphael Baysa',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Pixel Machinery',
+  },
+  {
+    Intern_ID: 20,
+    Name: 'Rawane Issa',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 14,
+    Name: 'Rudhra Raveendran',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Current BU Student',
+  },
+  {
+    Intern_ID: 46,
+    Name: 'Ruizhi Jiang',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 15,
+    Name: 'Ryan Walsh',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'HubSpot',
+  },
+  {
+    Intern_ID: 58,
+    Name: 'Sarah Leinicke',
+    Prior_Position: 'Software Engineer',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 57,
+    Name: 'Shreya Pandit',
+    Prior_Position: 'Software Engineer',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 34,
+    Name: 'Sidhartha Premkumar',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 13,
+    Name: 'Sreeja Keesara',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Goldman Sachs',
+  },
+  {
+    Intern_ID: 27,
+    Name: 'Stephanie Gomez',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Luminopia',
+  },
+  {
+    Intern_ID: 50,
+    Name: 'Taesung Yoon',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 44,
+    Name: 'Tanner Braun',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 9,
+    Name: 'Tim Lim',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Google',
+  },
+  {
+    Intern_ID: 37,
+    Name: 'Usama Ahsan',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 48,
+    Name: 'Xuan Hy Nguyen',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 42,
+    Name: 'Yi (Zack) Zhang',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+  {
+    Intern_ID: 33,
+    Name: 'Yiwen Gu',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: 'Ph.D. Program in Computer Science',
+  },
+  {
+    Intern_ID: 47,
+    Name: 'Yuwan Xiao',
+    Prior_Position: 'Software Engineer Intern',
+    Post_Internship: '',
+  },
+];
