--- src/utils/data/internInformation.tsx
@@ -1,313 +1,0 @@
-import { InternInformation } from '../../types/types';
-
-export const INTERN_ALUMNI: Array<InternInformation> = [
-  {
-    Intern_ID: 1,
-    Name: 'Chang Gao',
-    Cohort: '2016-SUM-SE,2017-SPR-SE',
-    Post_Internship: 'University of Virginia > Microsoft',
-  },
-  {
-    Intern_ID: 2,
-    Name: 'Aditi Dass',
-    Cohort: '2016-SUM-SE,2017-SUM-SE',
-    Post_Internship: 'State Street',
-  },
-  {
-    Intern_ID: 3,
-    Name: 'Ben Getchell',
-    Cohort: '2016-SUM-SE,2016-FAL-SE,2017-SPR-SE,2017-SUM-SE,2017-FAL-SE',
-    Post_Internship: 'Software & Application Innovation Lab',
-  },
-  {
-    Intern_ID: 4,
-    Name: 'Ekaterina Propokeva',
-    Cohort: '2016-SUM-SE',
-    Post_Internship: 'Amazon Web Services',
-  },
-  {
-    Intern_ID: 5,
-    Name: 'David Wang',
-    Cohort: '2016-SUM-SE,2017-SPR-SE',
-    Post_Internship: 'Amazon',
-  },
-  {
-    Intern_ID: 6,
-    Name: 'Jacqueline You',
-    Cohort:
-      '2016-SUM-SE,2016-FAL-SE,2017-SPR-SE,2017-SUM-SEF,2017-FAL-SEF,2018-SPR-SEF',
-    Post_Internship: 'Boston University School of Medicine',
-  },
-  {
-    Intern_ID: 7,
-    Name: 'Raphael Baysa',
-    Cohort: '2016-FAL-SE,2017-SPR-SE',
-    Post_Internship: 'Pixel Machinery',
-  },
-  {
-    Intern_ID: 8,
-    Name: 'Anna Goncharova',
-    Cohort: '2016-FAL-SE',
-    Post_Internship: 'Twitter',
-  },
-  {
-    Intern_ID: 9,
-    Name: 'Tim Lim',
-    Cohort: '2016-FAL-SE',
-    Post_Internship: 'Google',
-  },
-  {
-    Intern_ID: 10,
-    Name: 'Johnson Lam',
-    Cohort: '2016-FAL-SE',
-    Post_Internship: 'Soroco',
-  },
-  {
-    Intern_ID: 11,
-    Name: 'Ibrahim Shaikh',
-    Cohort: '2017-SPR-SE,2017-SUM-SE',
-    Post_Internship: 'ITG (needs confirmation)',
-  },
-  {
-    Intern_ID: 12,
-    Name: 'Linlan Chen',
-    Cohort: '2017-SPR-SE,2017-FAL-SE',
-    Post_Internship: 'Facebook',
-  },
-  {
-    Intern_ID: 13,
-    Name: 'Sreeja Keesara',
-    Cohort: '2017-SPR-SE',
-    Post_Internship: 'Goldman Sachs',
-  },
-  {
-    Intern_ID: 14,
-    Name: 'Rudhra Raveendran',
-    Cohort: '2017-SPR-SE,2017-FAL-SE,2018-SPR-SE',
-    Post_Internship: 'Current BU Student',
-  },
-  {
-    Intern_ID: 15,
-    Name: 'Ryan Walsh',
-    Cohort: '2017-SPR-SE',
-    Post_Internship: 'HubSpot',
-  },
-  {
-    Intern_ID: 16,
-    Name: 'Justin Chen',
-    Cohort: '2017-SUM-SE,2017-FAL-SE',
-    Post_Internship: 'Systems & Technology Research',
-  },
-  {
-    Intern_ID: 17,
-    Name: 'Kinan Dak Albab',
-    Cohort:
-      '2017-SUM-SEF,2017-FAL-SEF,2018-SPR-SEF,2018-SUM-SEF,2018-FAL-SEF,2019-SPR-SEF,2019-SUM-SEF',
-    Post_Internship: 'Current BU Ph.D. Student',
-  },
-  {
-    Intern_ID: 18,
-    Name: 'Chen Garber',
-    Cohort: '2017-SUM-SE',
-    Post_Internship: 'Facebook (needs confirmation)',
-  },
-  {
-    Intern_ID: 19,
-    Name: 'Qian Mei',
-    Cohort: '2017-SUM-GD,2017-FAL-GD,2018-SPR-GD',
-    Post_Internship: 'Growald Family Fund',
-  },
-  {
-    Intern_ID: 20,
-    Name: 'Rawane Issa',
-    Cohort:
-      '2017-SUM-SEF,2017-FAL-SEF,2018-SPR-SEF,2018-SUM-SEF,2018-FAL-SEF,2019-SPR-SEF,2019-SUM-SEF',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 21,
-    Name: 'Daniel Valentine',
-    Cohort: '2017-FAL-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 22,
-    Name: 'Esen Harris',
-    Cohort: '2017-FAL-SE,2018-SPR-SE,2018-SUM-SE,2018-FAL-SE',
-    Post_Internship: 'Accenture',
-  },
-  {
-    Intern_ID: 23,
-    Name: 'Anya Gilad',
-    Cohort: '2017-FAL-SE',
-    Post_Internship: 'Athenahealth',
-  },
-  {
-    Intern_ID: 27,
-    Name: 'Stephanie Gomez',
-    Cohort: '2018-SUM-SE',
-    Post_Internship: 'Luminopia',
-  },
-  {
-    Intern_ID: 28,
-    Name: 'Peter Flockhart',
-    Cohort: '2018-SUM-SE,2018-FAL-SE',
-    Post_Internship: 'Software & Application Innovation Lab',
-  },
-  {
-    Intern_ID: 29,
-    Name: 'Glenn Liem',
-    Cohort: '2018-SUM-SE,2021-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 30,
-    Name: 'Rachel Manzelli',
-    Cohort: '2018-SUM-SE,2018-FAL-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 31,
-    Name: 'Francis Zamora',
-    Cohort: '2018-SUM-SE,2018-FAL-SE',
-    Post_Internship: 'Twilio',
-  },
-  {
-    Intern_ID: 32,
-    Name: 'Nicolas Alhaddad',
-    Cohort: '2018-SUM-SEF,2018-FAL-SEF,2019-SPR-SEF,2019-SUM-SEF',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 33,
-    Name: 'Yiwen Gu',
-    Cohort: '2019-SPR-SE,2019-SUM-SE',
-    Post_Internship: 'Ph.D. Program in Computer Science',
-  },
-  {
-    Intern_ID: 34,
-    Name: 'Sidhartha Premkumar',
-    Cohort: '2019-SPR-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 35,
-    Name: 'David Shen',
-    Cohort: '2019-SPR-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 36,
-    Name: 'Megan Fantes',
-    Cohort: '2019-SUM-SE,2019-FAL-SE',
-    Post_Internship: 'iRobot',
-  },
-  {
-    Intern_ID: 37,
-    Name: 'Usama Ahsan',
-    Cohort: '2019-SUM-SE,2019-FAL-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 38,
-    Name: 'Efstathios Karatsiolis',
-    Cohort: '2019-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 39,
-    Name: 'FanonX Rogers',
-    Cohort: '2019-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 40,
-    Name: 'Ashni Shah',
-    Cohort: '2019-SUM-SE,2019-FAL-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 41,
-    Name: 'Christopher Trinh',
-    Cohort: '2019-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 42,
-    Name: 'Yi (Zack) Zhang',
-    Cohort: '2019-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 43,
-    Name: 'Marcella Hastings',
-    Cohort: '2019-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 44,
-    Name: 'Tanner Braun',
-    Cohort: '2020-SUM-SE,2021-SPR-SE,2020-FAL-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 45,
-    Name: 'Gagandeep Kang',
-    Cohort: '2020-SUM-SE,2020-FAL-SE',
-    Post_Internship: 'Amazon',
-  },
-  {
-    Intern_ID: 46,
-    Name: 'Ruizhi Jiang',
-    Cohort: '2020-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 47,
-    Name: 'Yuwan Xiao',
-    Cohort: '2020-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 48,
-    Name: 'Xuan Hy Nguyen',
-    Cohort: '2020-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 49,
-    Name: 'Raj Vipani',
-    Cohort: '2020-SUM-SE',
-    Post_Internship: 'Galatea Associates',
-  },
-  {
-    Intern_ID: 50,
-    Name: 'Taesung Yoon',
-    Cohort: '2021-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 51,
-    Name: 'Christopher Cho',
-    Cohort: '2021-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 52,
-    Name: 'Nicholas Goutermout',
-    Cohort: '2021-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 53,
-    Name: 'Nicholas Hall',
-    Cohort: '2021-SUM-SE',
-    Post_Internship: '',
-  },
-  {
-    Intern_ID: 54,
-    Name: 'Ammar Ahmad',
-    Cohort: '2021-SUM-SE',
-    Post_Internship: '',
-  },
-];
